refactor(responsestore): reduce complexity, split redis_store into files
Phase 6 of issue #2814: make the Phase 1-5 changes pass the repo's own
structural and complexity lint (make agent-lint), which nothing before
this point had actually run against this code.

Complexity fixes (golangci-lint cyclop/gocognit, no behavior change,
covered by the existing test suite):
- listIndexedResponseIDs split into rankInIndex (the ZRANK/ZREVRANK
  cursor lookup), resolveListWindow (the After/Before rank-window
  switch), and readIndexRange (the final ZRANGE/ZREVRANGE read).
- ensureConversationIndex split into backfillUnderLock (the
  acquired-lock branch) and awaitOrBackfillConversationIndex (the
  lock-wait-then-backfill branch).
- lazyBackfillConversationIndex split into scanConversationResponses
  (scan + deterministic sort, with its local struct promoted to the
  named discoveredConversationResponse type), setEmptyConversationIndexMarker,
  and indexBackfilledResponses (batched ZADD + TTL refresh + marker
  clear).

File-size fix (structure_check.py: warn at 400 lines, error at 800 -
redis_store.go had grown to 1745 lines and redis_store_test.go to
1371, both well past the error threshold): split both along the file's
own pre-existing "Response Store Methods" / "Conversation Store
Methods" / "Helper methods" section boundaries, matching how other
packages in this module (e.g. pkg/extproc's req_filter_rag_*.go) split
large files by feature area rather than carrying a single monolith.

  redis_store.go              -> construction/config/connection + key helpers
  redis_response_store.go     -> StoreResponse/GetResponse/UpdateResponse/
                                  DeleteResponse/GetConversationChain
  redis_list.go                -> ListResponsesByConversation + bounded
                                  windowed listing
  redis_conversation_store.go -> Create/Get/Update/DeleteConversation,
                                  ListConversations, AddResponseToConversation
  redis_conversation_index.go -> index/unindex, lazy migration, lock,
                                  Cluster-aware scan
  redis_helpers.go            -> storedConversationID, collectChainIDs,
                                  fetchResponsesPipelined

  redis_store_test.go              -> config/connection/key-isolation tests
  redis_test_helpers_test.go       -> shared test fixtures
  redis_response_store_test.go     -> StoreResponse/UpdateResponse tests
  redis_list_test.go               -> listing/backfill/cursor tests
  redis_conversation_store_test.go -> cascade delete/AddResponseToConversation tests

Purely organizational: no symbol renamed, no logic changed, same
package throughout. Verified with the full pkg/responsestore suite
(identical pass count before/after), go vet, go build, gofmt, and a
full `make test-semantic-router SKIP_REDIS_TESTS=false` module-wide run
(57 packages, 0 failures).
