+++ src/semantic-router/pkg/responsestore/redis_store.go
@@ -4,11 +4,8 @@ import (
 	"context"
 	"crypto/tls"
 	"crypto/x509"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"os"
-	"sort"
 	"strings"
 	"time"
 
@@ -16,7 +13,6 @@ import (
 	"sigs.k8s.io/yaml"
 
 	"github.com/vllm-project/semantic-router/src/semantic-router/pkg/observability/logging"
-	"github.com/vllm-project/semantic-router/src/semantic-router/pkg/responseapi"
 )
 
 // RedisStore implements the CombinedStore interface using Redis as the backend.
@@ -193,7 +189,6 @@ func NewRedisStore(config StoreConfig) (*RedisStore, error) {
 
 	return store, nil
 }
-
 func loadRedisStoreConfig(cfg RedisStoreConfig) (RedisStoreConfig, error) {
 	// If no external config, return inline config as-is
 	if cfg.ConfigPath == "" {
@@ -228,7 +223,6 @@ func loadRedisStoreConfig(cfg RedisStoreConfig) (RedisStoreConfig, error) {
 	// External file takes precedence
 	return fileCfg, nil
 }
-
 func validateRedisConfig(cfg RedisStoreConfig) error {
 	// Cluster mode validation
 	if cfg.ClusterMode {
@@ -266,7 +260,6 @@ func validateRedisConfig(cfg RedisStoreConfig) error {
 
 	return nil
 }
-
 func applyRedisConfigDefaults(cfg *RedisStoreConfig) {
 	if cfg.KeyPrefix == "" {
 		cfg.KeyPrefix = "sr:" // Base prefix only, types are added by constants
@@ -390,7 +383,6 @@ func (s *RedisStore) emptyConversationIndexMarkerKey(conversationID string) stri
 func (s *RedisStore) conversationIndexLockKey(conversationID string) string {
 	return s.buildKey(ConversationIndexLockKeyPrefix + conversationID)
 }
-
 func (s *RedisStore) CheckConnection(ctx context.Context) error {
 	if !s.enabled {
 		return fmt.Errorf("redis store is disabled")
@@ -404,7 +396,6 @@ func (s *RedisStore) CheckConnection(ctx context.Context) error {
 	logging.Debugf("RedisStore: connection check passed")
 	return nil
 }
-
 func (s *RedisStore) Close() error {
 	if s.client != nil {
 		logging.Infof("RedisStore: closing connection")
@@ -412,1276 +403,8 @@ func (s *RedisStore) Close() error {
 	}
 	return nil
 }
-
 func (s *RedisStore) IsEnabled() bool {
 	return s.enabled
 }
 
 // Response Store Methods
-
-func (s *RedisStore) StoreResponse(ctx context.Context, response *responseapi.StoredResponse) error {
-	if !s.enabled {
-		return ErrStoreDisabled
-	}
-	if response == nil || response.ID == "" {
-		return ErrInvalidInput
-	}
-
-	key := s.buildKey(ResponseKeyPrefix + response.ID)
-
-	data, err := json.Marshal(response)
-	if err != nil {
-		return fmt.Errorf("failed to serialize response: %w", err)
-	}
-
-	// Atomic existence check, and it lands the payload before the index entry, so
-	// a member always postdates its payload — what makes prune-on-missing safe.
-	stored, err := s.client.SetNX(ctx, key, data, s.ttl).Result()
-	if err != nil {
-		return fmt.Errorf("failed to store response in Redis: %w", err)
-	}
-	if !stored {
-		// The payload already here is the source of truth for whether this
-		// retry needs an index repair, not the caller's attempted request:
-		// verified inside repairExistingResponseIndex.
-		if response.ConversationID != "" {
-			if repairErr := s.repairExistingResponseIndex(ctx, response); repairErr != nil {
-				return repairErr
-			}
-		}
-		return ErrAlreadyExists
-	}
-
-	if response.ConversationID == "" {
-		return nil
-	}
-
-	if err := s.indexResponse(ctx, response.ConversationID, response.ID, response.CreatedAt); err != nil {
-		indexErr := fmt.Errorf("failed to index response in Redis: %w", err)
-
-		// Compare-delete rollback: never a blind DEL. Only removes the
-		// payload if it is still exactly what this call wrote, so a
-		// concurrent writer that stored a new value after this payload's TTL
-		// expired is never clobbered (the ABA race the blueprint calls out).
-		deleted, rollbackErr := s.compareDeleteResponsePayload(ctx, key, data)
-		if rollbackErr != nil {
-			return fmt.Errorf("%w (rollback failed: %v)", indexErr, rollbackErr)
-		}
-		if !deleted {
-			return fmt.Errorf("%w (payload changed before rollback, left in place)", indexErr)
-		}
-		return indexErr
-	}
-
-	return nil
-}
-
-// compareDeleteResponsePayload deletes key only if its current value equals
-// expected. Used to roll back a response payload after its index write
-// fails, without risking a blind DEL removing a value a concurrent writer
-// stored after the original payload expired on its TTL; also reused by
-// ensureConversationIndex to release the migration lock without releasing
-// one it doesn't hold (e.g. one that expired and was re-acquired).
-//
-// Single-key Lua script — touches only KEYS[1] — so it stays legal in Redis
-// Cluster; a two-key script spanning the response and index keys would not.
-func (s *RedisStore) compareDeleteResponsePayload(ctx context.Context, key string, expected []byte) (bool, error) {
-	res, err := compareDeleteScript.Run(ctx, s.client, []string{key}, expected).Result()
-	if err != nil {
-		return false, fmt.Errorf("failed to compare-delete response payload %s: %w", key, err)
-	}
-
-	deleted, ok := res.(int64)
-	if !ok {
-		return false, fmt.Errorf("unexpected compare-delete result type %T for %s", res, key)
-	}
-
-	return deleted > 0, nil
-}
-
-// repairExistingResponseIndex runs when StoreResponse's SETNX finds the
-// response ID already stored. It never trusts the caller's attempted
-// payload: the stored response is read back and is the only source of truth
-// for whether — and under which conversation — the index should be
-// repaired. A duplicate ID whose stored payload belongs to a different
-// conversation than the one the caller attempted must not poison that
-// conversation's index.
-func (s *RedisStore) repairExistingResponseIndex(ctx context.Context, attempted *responseapi.StoredResponse) error {
-	stored, err := s.GetResponse(ctx, attempted.ID)
-	if err != nil {
-		if errors.Is(err, ErrNotFound) {
-			// SETNX reported existence, but the payload is gone now (raced
-			// with a delete, or expired). Nothing to repair from; keep the
-			// duplicate contract.
-			return nil
-		}
-		return fmt.Errorf("failed to read stored response %s for index repair: %w", attempted.ID, err)
-	}
-
-	if stored.ConversationID == "" || stored.ConversationID != attempted.ConversationID {
-		// Either no index is expected, or the stored payload proves this
-		// duplicate belongs to a different conversation than attempted.
-		// Repairing the attempted conversation's index here would be
-		// indexing a response that conversation does not actually own.
-		return nil
-	}
-
-	if err := s.indexResponse(ctx, stored.ConversationID, stored.ID, stored.CreatedAt); err != nil {
-		return fmt.Errorf("response already exists but failed to repair conversation index: %w", err)
-	}
-
-	return nil
-}
-
-func (s *RedisStore) GetResponse(ctx context.Context, responseID string) (*responseapi.StoredResponse, error) {
-	if !s.enabled {
-		return nil, ErrStoreDisabled
-	}
-	if responseID == "" {
-		return nil, ErrInvalidInput
-	}
-
-	key := s.buildKey(ResponseKeyPrefix + responseID)
-
-	data, err := s.client.Get(ctx, key).Bytes()
-	if err != nil {
-		if errors.Is(err, redis.Nil) {
-			return nil, ErrNotFound
-		}
-		return nil, fmt.Errorf("failed to get response from Redis: %w", err)
-	}
-
-	var response responseapi.StoredResponse
-	if err := json.Unmarshal(data, &response); err != nil {
-		return nil, fmt.Errorf("failed to deserialize response: %w", err)
-	}
-
-	return &response, nil
-}
-
-// UpdateResponse replaces a response's payload and keeps its conversation
-// index in step: indexes under the new ConversationID (if any), then
-// best-effort unindexes the previous one if it changed. If the new index
-// write fails, restores the previous payload bytes and best-effort
-// reindexes the previous conversation before returning the error — an
-// update must never leave a payload pointing at a conversation whose index
-// was never actually written, matching the repairability blueprint §5 Phase
-// 5 asks for on top of StoreResponse's existing rollback/repair (Phase 2).
-//
-// Deliberate ordering deviation from the blueprint's UpdateResponse
-// pseudocode: this checks the new index write's outcome *before*
-// unindexing the previous conversation, not after. Unindexing first and
-// only then discovering the new write failed would leave a window where
-// neither the old nor the new conversation's index has the response, closed
-// here only by the subsequent best-effort reindex; checking first avoids
-// that window ever opening.
-func (s *RedisStore) UpdateResponse(ctx context.Context, response *responseapi.StoredResponse) error {
-	if !s.enabled {
-		return ErrStoreDisabled
-	}
-	if response == nil || response.ID == "" {
-		return ErrInvalidInput
-	}
-
-	key := s.buildKey(ResponseKeyPrefix + response.ID)
-
-	// Doubles as the existence check. Kept as raw bytes (not just the
-	// previously-parsed ConversationID storedConversationID would give) so a
-	// failed index write below can restore the exact previous payload.
-	previousData, err := s.client.Get(ctx, key).Bytes()
-	if err != nil {
-		if errors.Is(err, redis.Nil) {
-			return ErrNotFound
-		}
-		return fmt.Errorf("failed to check response existence: %w", err)
-	}
-
-	var previous responseapi.StoredResponse
-	previousConversationID := ""
-	if err := json.Unmarshal(previousData, &previous); err != nil {
-		// Not fatal: the update can still proceed, it just has nothing to
-		// unindex or restore from — same posture as storedConversationID.
-		logging.Warnf("RedisStore: failed to parse previous stored response %s during update: %v", response.ID, err)
-	} else {
-		previousConversationID = previous.ConversationID
-	}
-
-	data, err := json.Marshal(response)
-	if err != nil {
-		return fmt.Errorf("failed to serialize response: %w", err)
-	}
-
-	if err := s.client.Set(ctx, key, data, s.ttl).Err(); err != nil {
-		return fmt.Errorf("failed to update response in Redis: %w", err)
-	}
-
-	if response.ConversationID != "" {
-		if err := s.indexResponse(ctx, response.ConversationID, response.ID, response.CreatedAt); err != nil {
-			indexErr := fmt.Errorf("failed to index updated response in Redis: %w", err)
-
-			if restoreErr := s.client.Set(ctx, key, previousData, s.ttl).Err(); restoreErr != nil {
-				return fmt.Errorf("%w (failed to restore previous payload: %v)", indexErr, restoreErr)
-			}
-			if previousConversationID != "" {
-				if reindexErr := s.indexResponse(ctx, previousConversationID, response.ID, previous.CreatedAt); reindexErr != nil {
-					logging.Warnf("RedisStore: failed to reindex restored response %s under previous conversation %s after update rollback: %v",
-						response.ID, previousConversationID, reindexErr)
-				}
-			}
-			return indexErr
-		}
-	}
-
-	// The new index write (if any) is now confirmed in place, so the
-	// previous entry — if the conversation actually changed — is safe to
-	// drop. Best-effort: a stale leftover here is pruned by the next listing.
-	if previousConversationID != "" && previousConversationID != response.ConversationID {
-		if err := s.unindexResponse(ctx, previousConversationID, response.ID); err != nil {
-			logging.Warnf("RedisStore: failed to remove response %s from previous conversation %s index: %v",
-				response.ID, previousConversationID, err)
-		}
-	}
-
-	return nil
-}
-
-func (s *RedisStore) DeleteResponse(ctx context.Context, responseID string) error {
-	if !s.enabled {
-		return ErrStoreDisabled
-	}
-	if responseID == "" {
-		return ErrInvalidInput
-	}
-
-	key := s.buildKey(ResponseKeyPrefix + responseID)
-
-	// Needed to drop the index entry; also the existence check.
-	conversationID, err := s.storedConversationID(ctx, responseID)
-	if err != nil {
-		return err
-	}
-
-	deleted, err := s.client.Del(ctx, key).Result()
-	if err != nil {
-		return fmt.Errorf("failed to delete response from Redis: %w", err)
-	}
-	if deleted == 0 {
-		return ErrNotFound
-	}
-
-	// Best-effort: the payload delete above is the user-visible operation, and
-	// a stale index entry is pruned by the next listing that finds it missing.
-	if err := s.unindexResponse(ctx, conversationID, responseID); err != nil {
-		logging.Warnf("RedisStore: failed to remove response %s from conversation %s index: %v",
-			responseID, conversationID, err)
-	}
-
-	return nil
-}
-
-// GetConversationChain retrieves the full conversation chain for a response.
-// It follows the previous_response_id links backwards to build the complete history.
-func (s *RedisStore) GetConversationChain(ctx context.Context, responseID string) ([]*responseapi.StoredResponse, error) {
-	if !s.enabled {
-		return nil, ErrStoreDisabled
-	}
-	if responseID == "" {
-		return nil, ErrInvalidInput
-	}
-
-	// Phase 1: Collect response IDs by following the chain
-	responseIDs, err := s.collectChainIDs(ctx, responseID)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(responseIDs) == 0 {
-		return []*responseapi.StoredResponse{}, nil
-	}
-
-	// Phase 2: Fetch all responses using pipelining
-	chain, _, err := s.fetchResponsesPipelined(ctx, responseIDs)
-	if err != nil {
-		return nil, err
-	}
-
-	// Phase 3: Reverse chain to get chronological order (oldest first)
-	for i, j := 0, len(chain)-1; i < j; i, j = i+1, j-1 {
-		chain[i], chain[j] = chain[j], chain[i]
-	}
-
-	return chain, nil
-}
-
-// ListResponsesByConversation lists a conversation's responses via the
-// secondary index, at a cost proportional to the requested page rather than
-// the keyspace or even the conversation's full history (see
-// listIndexedResponseIDs).
-//
-// Read path: index exists → read the requested page from it. Otherwise
-// check the empty-conversation marker → nothing to do. Otherwise this may
-// be a pre-index conversation, an unknown ID, or a genuinely empty one that
-// has never been checked before: ensureConversationIndex resolves which,
-// running the one-time O(N) legacy scan at most once per conversation. See
-// blueprint §5 Phase 3 for the full state diagram this implements.
-//
-// Order/After/Before parity note: this implementation honors ListOptions.Order
-// (default "desc", newest first) and After/Before cursors, per the contract
-// documented on ListOptions in interface.go. MemoryStore does not — it
-// always returns insertion order regardless of these fields (see its own
-// doc comment). Bringing MemoryStore into line is out of scope for #2814;
-// callers that need a specific order from either backend today should not
-// assume Redis and MemoryStore agree on default order.
-func (s *RedisStore) ListResponsesByConversation(ctx context.Context, conversationID string, opts ListOptions) ([]*responseapi.StoredResponse, error) {
-	if !s.enabled {
-		return nil, ErrStoreDisabled
-	}
-	if conversationID == "" {
-		return nil, ErrInvalidInput
-	}
-
-	indexKey := s.conversationIndexKey(conversationID)
-
-	indexExists, err := s.client.Exists(ctx, indexKey).Result()
-	if err != nil {
-		return nil, fmt.Errorf("failed to check conversation index: %w", err)
-	}
-	if indexExists > 0 {
-		return s.listIndexedResponses(ctx, conversationID, opts)
-	}
-
-	emptyExists, err := s.client.Exists(ctx, s.emptyConversationIndexMarkerKey(conversationID)).Result()
-	if err != nil {
-		return nil, fmt.Errorf("failed to check conversation index empty marker: %w", err)
-	}
-	if emptyExists > 0 {
-		return nil, nil
-	}
-
-	// Neither exists: this conversation has never been resolved. Scan once
-	// (behind a migration lock, so concurrent readers of the same missing
-	// index don't all scan at once) and recheck.
-	if err := s.ensureConversationIndex(ctx, conversationID); err != nil {
-		return nil, err
-	}
-
-	indexExists, err = s.client.Exists(ctx, indexKey).Result()
-	if err != nil {
-		return nil, fmt.Errorf("failed to check conversation index: %w", err)
-	}
-	if indexExists > 0 {
-		return s.listIndexedResponses(ctx, conversationID, opts)
-	}
-
-	// The backfill (this call's own, or a concurrent one) found nothing and
-	// left the empty marker instead — or, if even that failed, the next call
-	// simply scans again. Either way there is nothing to return right now.
-	return nil, nil
-}
-
-// listIndexedResponses reads one page of a conversation's responses through
-// its already-confirmed-existing index: a bounded rank-window read
-// (listIndexedResponseIDs), not a full-index scan, so cost is proportional
-// to the page requested rather than the conversation's full history.
-//
-// Because only one page of IDs is read, pruning a stale or moved entry can
-// make this return fewer than the requested Limit even when more matching
-// responses exist further in the index. That is an accepted Phase 4
-// trade-off (blueprint §5 Phase 4): topping up short pages by re-reading
-// further windows would turn a bug fix into a pagination redesign.
-func (s *RedisStore) listIndexedResponses(ctx context.Context, conversationID string, opts ListOptions) ([]*responseapi.StoredResponse, error) {
-	responseIDs, err := s.listIndexedResponseIDs(ctx, conversationID, opts)
-	if err != nil {
-		return nil, err
-	}
-	if len(responseIDs) == 0 {
-		return nil, nil
-	}
-
-	fetched, missingIDs, err := s.fetchResponsesPipelined(ctx, responseIDs)
-	if err != nil {
-		return nil, err
-	}
-
-	// Payloads expire on their own TTL; their index entries do not. Pruning keeps
-	// a long-lived conversation's index from growing without bound. Best-effort:
-	// a failure here just costs the same prune again on the next listing.
-	if err := s.unindexResponse(ctx, conversationID, missingIDs...); err != nil {
-		logging.Warnf("RedisStore: failed to prune %d stale index entr(y/ies) from conversation %s: %v",
-			len(missingIDs), conversationID, err)
-	}
-
-	// Guard against an entry left behind by a response that moved
-	// conversation: prune it from this conversation's index too, not just
-	// filter it from this page, since it will never legitimately belong here.
-	responses := make([]*responseapi.StoredResponse, 0, len(fetched))
-	for _, response := range fetched {
-		if response.ConversationID == conversationID {
-			responses = append(responses, response)
-			continue
-		}
-		if err := s.unindexResponse(ctx, conversationID, response.ID); err != nil {
-			logging.Warnf("RedisStore: failed to prune response %s moved out of conversation %s: %v",
-				response.ID, conversationID, err)
-		}
-	}
-
-	return responses, nil
-}
-
-// normalizedListOptions is ListOptions after validation and defaulting:
-// Limit is always in [1, MaxListLimit], and Order is always exactly "asc"
-// or "desc".
-type normalizedListOptions struct {
-	Limit  int
-	Order  string
-	After  string
-	Before string
-}
-
-// normalizeResponseListOptions validates and defaults a caller's ListOptions
-// for indexed reads. Order defaults to "desc" (newest first), matching the
-// documented ListOptions.Order contract in interface.go and OpenAI's list
-// default — a contract neither store implementation actually honored before
-// this issue (both simply returned index/insertion order regardless of
-// Order). Rejects an unrecognized Order and rejects After and Before set
-// together, rather than silently picking one and ignoring the other.
-func normalizeResponseListOptions(opts ListOptions) (normalizedListOptions, error) {
-	if opts.After != "" && opts.Before != "" {
-		return normalizedListOptions{}, ErrInvalidInput
-	}
-
-	limit := opts.Limit
-	if limit <= 0 {
-		limit = DefaultListLimit
-	}
-	if limit > MaxListLimit {
-		limit = MaxListLimit
-	}
-
-	order := opts.Order
-	switch order {
-	case "":
-		order = "desc"
-	case "asc", "desc":
-		// already valid
-	default:
-		return normalizedListOptions{}, ErrInvalidInput
-	}
-
-	return normalizedListOptions{Limit: limit, Order: order, After: opts.After, Before: opts.Before}, nil
-}
-
-// listIndexedResponseIDs reads one bounded window of response IDs from a
-// conversation's index: at most normalizeResponseListOptions(opts).Limit
-// IDs, in the requested order, optionally positioned after/before a cursor
-// response ID — never a full ZRANGE 0 -1.
-//
-// Cursors are resolved via ZRANK (ascending order) or ZREVRANK (descending
-// order), i.e. rank in the order actually being read, and the window is
-// then read with the matching ZRANGE/ZREVRANGE. A cursor naming a response
-// ID that is not currently a member of the index (evicted, wrong
-// conversation, typo'd by the caller) yields an empty page rather than an
-// error: the same behavior as an ordinary page with nothing left to return.
-func (s *RedisStore) listIndexedResponseIDs(ctx context.Context, conversationID string, opts ListOptions) ([]string, error) {
-	normalized, err := normalizeResponseListOptions(opts)
-	if err != nil {
-		return nil, err
-	}
-
-	indexKey := s.conversationIndexKey(conversationID)
-	ascending := normalized.Order == "asc"
-
-	rank := func(member string) (int64, bool, error) {
-		var cmd *redis.IntCmd
-		if ascending {
-			cmd = s.client.ZRank(ctx, indexKey, member)
-		} else {
-			cmd = s.client.ZRevRank(ctx, indexKey, member)
-		}
-		r, err := cmd.Result()
-		if err != nil {
-			if errors.Is(err, redis.Nil) {
-				return 0, false, nil
-			}
-			return 0, false, fmt.Errorf("failed to rank conversation index cursor %s: %w", member, err)
-		}
-		return r, true, nil
-	}
-
-	limit := int64(normalized.Limit)
-	var start, end int64
-
-	switch {
-	case normalized.After != "":
-		r, found, err := rank(normalized.After)
-		if err != nil {
-			return nil, err
-		}
-		if !found {
-			return nil, nil
-		}
-		start = r + 1
-		end = start + limit - 1
-	case normalized.Before != "":
-		r, found, err := rank(normalized.Before)
-		if err != nil {
-			return nil, err
-		}
-		if !found {
-			return nil, nil
-		}
-		end = r - 1
-		start = end - limit + 1
-		if start < 0 {
-			start = 0
-		}
-	default:
-		start = 0
-		end = limit - 1
-	}
-
-	if end < start {
-		return nil, nil
-	}
-
-	var idsCmd *redis.StringSliceCmd
-	if ascending {
-		idsCmd = s.client.ZRange(ctx, indexKey, start, end)
-	} else {
-		idsCmd = s.client.ZRevRange(ctx, indexKey, start, end)
-	}
-	ids, err := idsCmd.Result()
-	if err != nil {
-		return nil, fmt.Errorf("failed to read conversation index window: %w", err)
-	}
-
-	return ids, nil
-}
-
-// Conversation Store Methods
-
-func (s *RedisStore) CreateConversation(ctx context.Context, conversation *responseapi.StoredConversation) error {
-	if !s.enabled {
-		return ErrStoreDisabled
-	}
-	if conversation == nil || conversation.ID == "" {
-		return ErrInvalidInput
-	}
-
-	key := s.buildKey(ConversationKeyPrefix + conversation.ID)
-
-	exists, err := s.client.Exists(ctx, key).Result()
-	if err != nil {
-		return fmt.Errorf("failed to check conversation existence: %w", err)
-	}
-	if exists > 0 {
-		return ErrAlreadyExists
-	}
-
-	data, err := json.Marshal(conversation)
-	if err != nil {
-		return fmt.Errorf("failed to serialize conversation: %w", err)
-	}
-
-	if err := s.client.Set(ctx, key, data, s.ttl).Err(); err != nil {
-		return fmt.Errorf("failed to store conversation in Redis: %w", err)
-	}
-
-	return nil
-}
-
-func (s *RedisStore) GetConversation(ctx context.Context, conversationID string) (*responseapi.StoredConversation, error) {
-	if !s.enabled {
-		return nil, ErrStoreDisabled
-	}
-	if conversationID == "" {
-		return nil, ErrInvalidInput
-	}
-
-	key := s.buildKey(ConversationKeyPrefix + conversationID)
-
-	data, err := s.client.Get(ctx, key).Bytes()
-	if err != nil {
-		if errors.Is(err, redis.Nil) {
-			return nil, ErrNotFound
-		}
-		return nil, fmt.Errorf("failed to get conversation from Redis: %w", err)
-	}
-
-	var conversation responseapi.StoredConversation
-	if err := json.Unmarshal(data, &conversation); err != nil {
-		return nil, fmt.Errorf("failed to deserialize conversation: %w", err)
-	}
-
-	return &conversation, nil
-}
-
-func (s *RedisStore) UpdateConversation(ctx context.Context, conversation *responseapi.StoredConversation) error {
-	if !s.enabled {
-		return ErrStoreDisabled
-	}
-	if conversation == nil || conversation.ID == "" {
-		return ErrInvalidInput
-	}
-
-	key := s.buildKey(ConversationKeyPrefix + conversation.ID)
-
-	exists, err := s.client.Exists(ctx, key).Result()
-	if err != nil {
-		return fmt.Errorf("failed to check conversation existence: %w", err)
-	}
-	if exists == 0 {
-		return ErrNotFound
-	}
-
-	data, err := json.Marshal(conversation)
-	if err != nil {
-		return fmt.Errorf("failed to serialize conversation: %w", err)
-	}
-
-	if err := s.client.Set(ctx, key, data, s.ttl).Err(); err != nil {
-		return fmt.Errorf("failed to update conversation in Redis: %w", err)
-	}
-
-	return nil
-}
-
-// DeleteConversation deletes a conversation and, when deleteResponses is
-// set, cascades to its indexed responses first. The conversation record is
-// deleted last on purpose: if the cascade fails partway, the conversation
-// key survives as the retry anchor instead of leaving orphaned response
-// payloads with no record pointing at them (blueprint §5 Phase 5).
-func (s *RedisStore) DeleteConversation(ctx context.Context, conversationID string, deleteResponses bool) error {
-	if !s.enabled {
-		return ErrStoreDisabled
-	}
-	if conversationID == "" {
-		return ErrInvalidInput
-	}
-
-	convKey := s.buildKey(ConversationKeyPrefix + conversationID)
-
-	exists, err := s.client.Exists(ctx, convKey).Result()
-	if err != nil {
-		return fmt.Errorf("failed to check conversation existence: %w", err)
-	}
-	if exists == 0 {
-		return ErrNotFound
-	}
-
-	if deleteResponses {
-		if err := s.deleteConversationResponses(ctx, conversationID); err != nil {
-			return err
-		}
-	}
-
-	if err := s.client.Del(ctx, convKey).Err(); err != nil {
-		return fmt.Errorf("failed to delete conversation from Redis: %w", err)
-	}
-
-	return nil
-}
-
-// deleteConversationResponses removes a conversation's responses and its
-// index/marker, in bounded batches of redisDeleteBatchSize read from the
-// front of the index — never one ZRANGE 0 -1 or one pipeline sized to the
-// whole conversation. Reads the index directly rather than through
-// ListResponsesByConversation, which would additionally cap the cascade at
-// a single page.
-//
-// Each iteration reads rank 0..redisDeleteBatchSize-1 again (not an
-// offsetting range): the ZREM at the end of the previous iteration already
-// removed those members, so the next read naturally advances. If a batch's
-// payload deletes or index removal fails, this returns the error and stops
-// rather than silently logging and reporting success — per blueprint §5
-// Phase 5, the caller needs to know the cascade was only partially applied.
-// The failure is safely retryable: undeleted response keys are unaffected,
-// and a Del of an already-deleted key is a no-op, not an error.
-func (s *RedisStore) deleteConversationResponses(ctx context.Context, conversationID string) error {
-	indexKey := s.conversationIndexKey(conversationID)
-
-	for {
-		responseIDs, err := s.client.ZRange(ctx, indexKey, 0, redisDeleteBatchSize-1).Result()
-		if err != nil {
-			return fmt.Errorf("failed to list responses for deletion: %w", err)
-		}
-		if len(responseIDs) == 0 {
-			break
-		}
-
-		if s.deleteResponseBatchOverride != nil {
-			if err := s.deleteResponseBatchOverride(ctx, responseIDs); err != nil {
-				return fmt.Errorf("failed to delete %d response(s) of conversation %s: %w",
-					len(responseIDs), conversationID, err)
-			}
-		} else {
-			pipe := s.client.Pipeline()
-			for _, responseID := range responseIDs {
-				pipe.Del(ctx, s.buildKey(ResponseKeyPrefix+responseID))
-			}
-			if _, err := pipe.Exec(ctx); err != nil {
-				return fmt.Errorf("failed to delete %d response(s) of conversation %s: %w",
-					len(responseIDs), conversationID, err)
-			}
-		}
-
-		if err := s.unindexResponse(ctx, conversationID, responseIDs...); err != nil {
-			return fmt.Errorf("failed to remove deleted responses from conversation %s index: %w", conversationID, err)
-		}
-	}
-
-	// Single-key deletes, never combined with each other or with the
-	// response keys above.
-	if err := s.client.Del(ctx, indexKey).Err(); err != nil {
-		return fmt.Errorf("failed to delete conversation index for %s: %w", conversationID, err)
-	}
-	if err := s.client.Del(ctx, s.emptyConversationIndexMarkerKey(conversationID)).Err(); err != nil {
-		return fmt.Errorf("failed to delete conversation index empty marker for %s: %w", conversationID, err)
-	}
-
-	return nil
-}
-
-func (s *RedisStore) ListConversations(ctx context.Context, opts ListOptions) ([]*responseapi.StoredConversation, error) {
-	if !s.enabled {
-		return nil, ErrStoreDisabled
-	}
-
-	pattern := s.buildKey(ConversationKeyPrefix + "*")
-	var conversations []*responseapi.StoredConversation
-
-	iter := s.client.Scan(ctx, 0, pattern, 0).Iterator()
-	for iter.Next(ctx) {
-		key := iter.Val()
-
-		data, err := s.client.Get(ctx, key).Bytes()
-		if err != nil {
-			continue
-		}
-
-		var conversation responseapi.StoredConversation
-		if err := json.Unmarshal(data, &conversation); err != nil {
-			continue
-		}
-
-		conversations = append(conversations, &conversation)
-	}
-
-	if err := iter.Err(); err != nil {
-		return nil, fmt.Errorf("failed to scan conversations: %w", err)
-	}
-
-	// Apply list options (limit, pagination)
-	conversations = ApplyConvListOptions(conversations, opts)
-
-	return conversations, nil
-}
-
-// AddResponseToConversation indexes an already-stored response under a
-// conversation, verified against the response's own stored ConversationID —
-// never against the caller's say-so alone, so this can never create an
-// index entry for a conversation the response doesn't actually belong to.
-//
-// Deliberately narrow (blueprint §3.7/§7): a response with no stored
-// ConversationID, or one that belongs to a different conversation than
-// conversationID, returns ErrInvalidInput rather than silently adopting the
-// caller's conversationID or rewriting the stored payload's membership.
-// Widening this to actually reassign a response's conversation belongs to
-// the broader Conversations API semantics in #2999, not this lookup fix —
-// today's only real caller path is StoreResponse indexing directly from
-// StoredResponse.ConversationID, so this method mainly exists to satisfy
-// the ConversationStore interface usefully rather than as a no-op.
-func (s *RedisStore) AddResponseToConversation(ctx context.Context, conversationID, responseID string) error {
-	if !s.enabled {
-		return ErrStoreDisabled
-	}
-	if conversationID == "" || responseID == "" {
-		return ErrInvalidInput
-	}
-
-	stored, err := s.GetResponse(ctx, responseID)
-	if err != nil {
-		return err
-	}
-
-	if stored.ConversationID == "" || stored.ConversationID != conversationID {
-		return ErrInvalidInput
-	}
-
-	if err := s.indexResponse(ctx, conversationID, responseID, stored.CreatedAt); err != nil {
-		return fmt.Errorf("failed to index response in Redis: %w", err)
-	}
-
-	return nil
-}
-
-// Helper methods
-
-// indexResponse adds a response to its conversation index, scored by
-// created_at; refreshes the index TTL; and clears the empty-conversation
-// marker so a newly indexed write is never hidden behind a stale "scanned,
-// found nothing" marker (checked here best-effort — a marker left behind by
-// this call failing is harmless, because every read checks the index before
-// the marker).
-//
-// Returns an error instead of swallowing it: the payload this indexes is
-// already durable by the time this runs (StoreResponse writes it first), so
-// the caller — not this helper — must decide what an index failure means:
-// StoreResponse rolls the payload back, UpdateResponse restores the previous
-// payload, DeleteResponse and lazy backfill may choose to log and continue.
-func (s *RedisStore) indexResponse(ctx context.Context, conversationID, responseID string, createdAt int64) error {
-	if conversationID == "" || responseID == "" {
-		return nil
-	}
-	if s.indexResponseOverride != nil {
-		return s.indexResponseOverride(ctx, conversationID, responseID, createdAt)
-	}
-
-	indexKey := s.conversationIndexKey(conversationID)
-
-	pipe := s.client.Pipeline()
-	pipe.ZAdd(ctx, indexKey, redis.Z{Score: float64(createdAt), Member: responseID})
-	if s.ttl > 0 {
-		// Outlive the newest member. Guarded: EXPIRE with 0 deletes the key.
-		pipe.Expire(ctx, indexKey, s.ttl)
-	}
-	// Single-key DEL, never combined with another key: Cluster safe. A no-op
-	// (returns 0, not an error) when no marker exists.
-	pipe.Del(ctx, s.emptyConversationIndexMarkerKey(conversationID))
-
-	if _, err := pipe.Exec(ctx); err != nil {
-		return fmt.Errorf("failed to index response %s in conversation %s: %w", responseID, conversationID, err)
-	}
-
-	return nil
-}
-
-// unindexResponse drops response IDs from a conversation index. ZREM is
-// variadic but touches only one key (the zset), all members belong to the
-// same conversation index, so it stays Cluster safe.
-func (s *RedisStore) unindexResponse(ctx context.Context, conversationID string, responseIDs ...string) error {
-	if conversationID == "" || len(responseIDs) == 0 {
-		return nil
-	}
-
-	members := make([]interface{}, len(responseIDs))
-	for i, responseID := range responseIDs {
-		members[i] = responseID
-	}
-
-	if err := s.client.ZRem(ctx, s.conversationIndexKey(conversationID), members...).Err(); err != nil {
-		return fmt.Errorf("failed to remove %d response(s) from conversation %s index: %w", len(responseIDs), conversationID, err)
-	}
-
-	return nil
-}
-
-// legacyBackfillResult reports how many responses a lazy backfill scan found
-// and indexed for a conversation, mainly for logging.
-type legacyBackfillResult struct {
-	// Found is the number of responses discovered belonging to the scanned
-	// conversation. Zero means the empty marker was set instead.
-	Found int
-}
-
-// indexOrMarkerExists reports whether a conversation's index or its empty
-// marker exists, via two single-key EXISTS calls. Deliberately not a
-// variadic Exists(ctx, indexKey, markerKey): the two keys have different
-// prefixes and are not guaranteed to share a Cluster hash slot.
-func (s *RedisStore) indexOrMarkerExists(ctx context.Context, conversationID string) (bool, error) {
-	indexExists, err := s.client.Exists(ctx, s.conversationIndexKey(conversationID)).Result()
-	if err != nil {
-		return false, fmt.Errorf("failed to check conversation index: %w", err)
-	}
-	if indexExists > 0 {
-		return true, nil
-	}
-
-	emptyExists, err := s.client.Exists(ctx, s.emptyConversationIndexMarkerKey(conversationID)).Result()
-	if err != nil {
-		return false, fmt.Errorf("failed to check conversation index empty marker: %w", err)
-	}
-
-	return emptyExists > 0, nil
-}
-
-// ensureConversationIndex guarantees that, barring a concurrent delete of
-// both immediately afterward, either the conversation's index or its empty
-// marker exists once this returns without error. It runs the O(N) legacy
-// scan (lazyBackfillConversationIndex) at most once per conversation per
-// marker/index lifetime.
-//
-// A short migration lock (SET NX PX conversationIndexMigrationLockTTL)
-// keeps concurrent readers of the same missing index from all scanning at
-// once, but holding it is an optimization, not a correctness dependency: a
-// reader that cannot acquire it backs off briefly, rechecks, and — if the
-// holder appears to have died mid-scan (pod restart, deadline) rather than
-// finished — runs the scan itself anyway. Every path converges on the same
-// additive, idempotent backfill.
-func (s *RedisStore) ensureConversationIndex(ctx context.Context, conversationID string) error {
-	lockKey := s.conversationIndexLockKey(conversationID)
-	token := []byte(fmt.Sprintf("%d:%d", time.Now().UnixNano(), os.Getpid()))
-
-	acquired, err := s.client.SetNX(ctx, lockKey, token, conversationIndexMigrationLockTTL).Result()
-	if err != nil {
-		return fmt.Errorf("failed to acquire conversation index migration lock: %w", err)
-	}
-
-	if acquired {
-		defer func() {
-			// Reused single-key compare-delete: never releases a lock this
-			// call didn't acquire (e.g. one that already expired and was
-			// re-acquired by someone else).
-			if _, releaseErr := s.compareDeleteResponsePayload(ctx, lockKey, token); releaseErr != nil {
-				logging.Debugf("RedisStore: failed to release conversation index migration lock for %s: %v",
-					conversationID, releaseErr)
-			}
-		}()
-
-		// Recheck under the lock: a writer, or a backfill that started just
-		// before this one acquired it, may have already resolved this
-		// conversation.
-		exists, existsErr := s.indexOrMarkerExists(ctx, conversationID)
-		if existsErr != nil {
-			return existsErr
-		}
-		if exists {
-			return nil
-		}
-
-		result, backfillErr := s.lazyBackfillConversationIndex(ctx, conversationID)
-		if backfillErr != nil {
-			return backfillErr
-		}
-		logging.Debugf("RedisStore: lazy-backfilled conversation %s index with %d response(s)",
-			conversationID, result.Found)
-		return nil
-	}
-
-	// Someone else holds the lock. Back off briefly and recheck; if the
-	// index/marker still hasn't appeared once the bounded wait is over, the
-	// holder may have died mid-scan — run the scan locally rather than block
-	// indefinitely. Correctness wins over avoiding duplicate work, which is
-	// the lock's only job.
-	const (
-		lockWaitAttempts = 5
-		lockWaitDelay    = 100 * time.Millisecond
-	)
-	for i := 0; i < lockWaitAttempts; i++ {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-time.After(lockWaitDelay):
-		}
-
-		exists, existsErr := s.indexOrMarkerExists(ctx, conversationID)
-		if existsErr != nil {
-			return existsErr
-		}
-		if exists {
-			return nil
-		}
-	}
-
-	result, err := s.lazyBackfillConversationIndex(ctx, conversationID)
-	if err != nil {
-		return err
-	}
-	logging.Debugf("RedisStore: lazy-backfilled conversation %s index with %d response(s) after lock wait",
-		conversationID, result.Found)
-	return nil
-}
-
-// lazyBackfillConversationIndex performs the one-time O(N) scan that makes a
-// pre-index conversation's responses discoverable: it walks every response
-// payload once, keeps the ones matching conversationID, and either indexes
-// them or — if none match — sets the empty marker so the next read does not
-// scan again.
-//
-// Idempotent and additive: it only ZADDs discovered members and never
-// deletes, so a concurrently indexed write (e.g. a StoreResponse racing this
-// scan) is never undone by it, no matter which finishes first.
-func (s *RedisStore) lazyBackfillConversationIndex(ctx context.Context, conversationID string) (legacyBackfillResult, error) {
-	if s.lazyBackfillPreScanHook != nil {
-		s.lazyBackfillPreScanHook()
-	}
-
-	type discovered struct {
-		id        string
-		createdAt int64
-	}
-	var found []discovered
-
-	err := s.scanResponsePayloads(ctx, func(batch []*responseapi.StoredResponse) error {
-		for _, response := range batch {
-			if response.ConversationID == conversationID {
-				found = append(found, discovered{id: response.ID, createdAt: response.CreatedAt})
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return legacyBackfillResult{}, fmt.Errorf("failed to backfill conversation index: %w", err)
-	}
-
-	if len(found) == 0 {
-		if err := s.client.Set(ctx, s.emptyConversationIndexMarkerKey(conversationID), "v1", s.ttl).Err(); err != nil {
-			// Correctness is preserved either way: the next read scans
-			// again rather than trusting a marker that isn't there.
-			logging.Debugf("RedisStore: failed to set empty conversation index marker for %s: %v",
-				conversationID, err)
-		}
-		return legacyBackfillResult{Found: 0}, nil
-	}
-
-	// Deterministic order: primary by created_at, tie-broken by response ID
-	// (blueprint §3.6) — never fabricate sub-second ordering from wall clock
-	// time during repair.
-	sort.Slice(found, func(i, j int) bool {
-		if found[i].createdAt != found[j].createdAt {
-			return found[i].createdAt < found[j].createdAt
-		}
-		return found[i].id < found[j].id
-	})
-
-	indexKey := s.conversationIndexKey(conversationID)
-	for start := 0; start < len(found); start += redisBackfillBatchSize {
-		end := start + redisBackfillBatchSize
-		if end > len(found) {
-			end = len(found)
-		}
-
-		members := make([]redis.Z, end-start)
-		for i, d := range found[start:end] {
-			members[i] = redis.Z{Score: float64(d.createdAt), Member: d.id}
-		}
-		if err := s.client.ZAdd(ctx, indexKey, members...).Err(); err != nil {
-			return legacyBackfillResult{}, fmt.Errorf("failed to backfill conversation index: %w", err)
-		}
-	}
-
-	if s.ttl > 0 {
-		if err := s.client.Expire(ctx, indexKey, s.ttl).Err(); err != nil {
-			logging.Warnf("RedisStore: failed to refresh TTL on backfilled conversation index %s: %v",
-				conversationID, err)
-		}
-	}
-
-	// Best-effort: the index now exists, and every read checks it before the
-	// marker, so a marker left behind here is harmless even if this DEL fails.
-	if err := s.client.Del(ctx, s.emptyConversationIndexMarkerKey(conversationID)).Err(); err != nil {
-		logging.Debugf("RedisStore: failed to clear empty conversation index marker for %s: %v",
-			conversationID, err)
-	}
-
-	return legacyBackfillResult{Found: len(found)}, nil
-}
-
-// scanResponsePayloads walks every response payload key exactly once,
-// decoding each into a StoredResponse and delivering them to visit in
-// bounded batches (redisBackfillBatchSize).
-//
-// Cluster-aware: a single Redis Cluster node's keyspace only holds the slots
-// assigned to it, so in Cluster mode this scans every master via
-// ForEachMaster. Standalone mode scans the one client directly.
-//
-// Used only by lazy legacy backfill — this is the O(N) operation the index
-// exists to avoid on the hot read path.
-func (s *RedisStore) scanResponsePayloads(ctx context.Context, visit func(batch []*responseapi.StoredResponse) error) error {
-	s.scanInvocations++
-
-	pattern := s.buildKey(ResponseKeyPrefix + "*")
-
-	scanNode := func(ctx context.Context, client redis.UniversalClient) error {
-		var keys []string
-		flush := func() error {
-			if len(keys) == 0 {
-				return nil
-			}
-			batch := s.getResponsesPipelined(ctx, client, keys)
-			keys = keys[:0]
-			if len(batch) == 0 {
-				return nil
-			}
-			return visit(batch)
-		}
-
-		iter := client.Scan(ctx, 0, pattern, redisScanCount).Iterator()
-		for iter.Next(ctx) {
-			keys = append(keys, iter.Val())
-			if len(keys) >= redisBackfillBatchSize {
-				if err := flush(); err != nil {
-					return err
-				}
-			}
-		}
-		if err := iter.Err(); err != nil {
-			return fmt.Errorf("failed to scan response keys: %w", err)
-		}
-
-		return flush()
-	}
-
-	if clusterClient, ok := s.client.(*redis.ClusterClient); ok {
-		return clusterClient.ForEachMaster(ctx, func(ctx context.Context, master *redis.Client) error {
-			return scanNode(ctx, master)
-		})
-	}
-
-	return scanNode(ctx, s.client)
-}
-
-// getResponsesPipelined GETs and decodes payloads for a batch of already-
-// prefixed keys in one round trip against the given client. Malformed or
-// missing payloads are skipped with a log line rather than failing the
-// batch: a legacy scan must make forward progress even if one record is
-// corrupt or expired mid-scan.
-func (s *RedisStore) getResponsesPipelined(ctx context.Context, client redis.UniversalClient, keys []string) []*responseapi.StoredResponse {
-	if len(keys) == 0 {
-		return nil
-	}
-
-	pipe := client.Pipeline()
-	cmds := make([]*redis.StringCmd, len(keys))
-	for i, key := range keys {
-		cmds[i] = pipe.Get(ctx, key)
-	}
-	if _, err := pipe.Exec(ctx); err != nil && !errors.Is(err, redis.Nil) {
-		logging.Debugf("RedisStore: scan pipeline execution completed with some errors: %v", err)
-	}
-
-	responses := make([]*responseapi.StoredResponse, 0, len(keys))
-	for i, cmd := range cmds {
-		data, err := cmd.Bytes()
-		if err != nil {
-			if !errors.Is(err, redis.Nil) {
-				logging.Warnf("RedisStore: failed to get response at key %s during scan: %v", keys[i], err)
-			}
-			continue
-		}
-
-		var response responseapi.StoredResponse
-		if err := json.Unmarshal(data, &response); err != nil {
-			logging.Warnf("RedisStore: failed to parse response at key %s during scan: %v", keys[i], err)
-			continue
-		}
-		if response.ID == "" {
-			continue
-		}
-
-		responses = append(responses, &response)
-	}
-
-	return responses
-}
-
-// storedConversationID reports a response's current conversation so update and
-// delete can repair the index, and returns ErrNotFound when it is absent. An
-// unreadable payload is not fatal, it only costs the old index entry.
-func (s *RedisStore) storedConversationID(ctx context.Context, responseID string) (string, error) {
-	data, err := s.client.Get(ctx, s.buildKey(ResponseKeyPrefix+responseID)).Bytes()
-	if err != nil {
-		if errors.Is(err, redis.Nil) {
-			return "", ErrNotFound
-		}
-		return "", fmt.Errorf("failed to check response existence: %w", err)
-	}
-
-	var stored responseapi.StoredResponse
-	if err := json.Unmarshal(data, &stored); err != nil {
-		logging.Warnf("RedisStore: failed to parse stored response %s while updating its conversation index: %v",
-			responseID, err)
-		return "", nil
-	}
-
-	return stored.ConversationID, nil
-}
-
-func (s *RedisStore) collectChainIDs(ctx context.Context, startID string) ([]string, error) {
-	var responseIDs []string
-	currentID := startID
-	visited := make(map[string]bool)
-
-	// Maximum chain length to prevent infinite loops
-	const maxChainLength = 1000
-
-	for currentID != "" && len(responseIDs) < maxChainLength {
-		// Prevent circular references
-		if visited[currentID] {
-			logging.Warnf("RedisStore: circular reference detected at %s", currentID)
-			break
-		}
-		visited[currentID] = true
-
-		responseIDs = append(responseIDs, currentID)
-
-		response, err := s.GetResponse(ctx, currentID)
-		if err != nil {
-			if errors.Is(err, ErrNotFound) {
-				// If this is the first response (start of chain), return error
-				if len(responseIDs) == 1 {
-					return nil, ErrNotFound
-				}
-				// Otherwise, just break - the chain ended early
-				logging.Warnf("RedisStore: response %s not found in chain", currentID)
-				break
-			}
-			return nil, fmt.Errorf("failed to fetch response %s: %w", currentID, err)
-		}
-
-		currentID = response.PreviousResponseID
-	}
-
-	return responseIDs, nil
-}
-
-// fetchResponsesPipelined loads response IDs in one round trip, also returning
-// the IDs whose payload is gone so index-driven callers can prune them.
-func (s *RedisStore) fetchResponsesPipelined(ctx context.Context, responseIDs []string) ([]*responseapi.StoredResponse, []string, error) {
-	if len(responseIDs) == 0 {
-		return []*responseapi.StoredResponse{}, nil, nil
-	}
-
-	pipe := s.client.Pipeline()
-
-	cmds := make([]*redis.StringCmd, len(responseIDs))
-	for i, id := range responseIDs {
-		key := s.buildKey(ResponseKeyPrefix + id)
-		cmds[i] = pipe.Get(ctx, key)
-	}
-
-	_, err := pipe.Exec(ctx)
-	if err != nil && !errors.Is(err, redis.Nil) {
-		// Some commands might fail, but we continue to process successful ones
-		logging.Debugf("RedisStore: pipeline execution completed with some errors: %v", err)
-	}
-
-	// Process results
-	var (
-		found      []*responseapi.StoredResponse
-		missingIDs []string
-	)
-	for i, cmd := range cmds {
-		data, err := cmd.Bytes()
-		if err != nil {
-			if errors.Is(err, redis.Nil) {
-				logging.Warnf("RedisStore: response %s not found (may have expired)", responseIDs[i])
-				missingIDs = append(missingIDs, responseIDs[i])
-				continue
-			}
-			logging.Warnf("RedisStore: failed to get response %s: %v", responseIDs[i], err)
-			continue
-		}
-
-		var response responseapi.StoredResponse
-		if err := json.Unmarshal(data, &response); err != nil {
-			logging.Warnf("RedisStore: failed to parse response %s: %v", responseIDs[i], err)
-			continue
-		}
-
-		found = append(found, &response)
-	}
-
-	return found, missingIDs, nil
-}
