+++ src/semantic-router/pkg/responsestore/redis_store_test.go
@@ -2,14 +2,10 @@ package responsestore
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -507,864 +503,4 @@ func TestRedisConversationIndexKeyIsolation(t *testing.T) {
 		"empty marker key %q must not be matched by the %q* index scan pattern", emptyMarkerKey, indexScanPrefix)
 	assert.Falsef(t, strings.HasPrefix(lockKey, indexScanPrefix),
 		"lock key %q must not be matched by the %q* index scan pattern", lockKey, indexScanPrefix)
-}
-
-// conversationIndexMembers reads the index directly, so tests can assert on it
-// and not only on what a listing happens to return.
-func conversationIndexMembers(t *testing.T, store *RedisStore, conversationID string) []string {
-	t.Helper()
-
-	members, err := store.client.ZRange(context.Background(), store.conversationIndexKey(conversationID), 0, -1).Result()
-	require.NoError(t, err)
-	return members
-}
-
-// newConversationIndexStore scopes the store to a key prefix unique to this run,
-// so it cannot collide with the other suites sharing DB 0. Skips without Redis.
-func newConversationIndexStore(t *testing.T) *RedisStore {
-	t.Helper()
-
-	cfg := StoreConfig{
-		Enabled:     true,
-		TTLSeconds:  300,
-		BackendType: RedisStoreType,
-		Redis: RedisStoreConfig{
-			Address:   "localhost:6379",
-			DB:        0,
-			KeyPrefix: fmt.Sprintf("srtest:%d:", time.Now().UnixNano()),
-		},
-	}
-
-	store, err := NewRedisStore(cfg)
-	if err != nil {
-		t.Skipf("Redis not available: %v", err)
-	}
-
-	t.Cleanup(func() {
-		ctx := context.Background()
-		iter := store.client.Scan(ctx, 0, store.buildKey("*"), 0).Iterator()
-		for iter.Next(ctx) {
-			store.client.Del(ctx, iter.Val())
-		}
-		_ = store.Close()
-	})
-
-	return store
-}
-
-// TestRedisConversationIndexListing covers the index-backed read path: what
-// it returns, in what order, and how it converges when index and payloads
-// disagree. Subtests that care about relative order pass Order: "asc"
-// explicitly — the default order is "desc" (newest first, blueprint §3.6,
-// matching the ListOptions.Order contract in interface.go), which
-// TestRedisListDefaultOrderIsDescending covers directly.
-func TestRedisConversationIndexListing(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	now := time.Now().Unix()
-	storeResponse := func(t *testing.T, id, convID string, createdAt int64) {
-		t.Helper()
-		require.NoError(t, store.StoreResponse(ctx, &responseapi.StoredResponse{
-			ID:             id,
-			ConversationID: convID,
-			Status:         "completed",
-			CreatedAt:      createdAt,
-			Output:         []responseapi.OutputItem{{ID: "item_" + id}},
-		}))
-	}
-
-	t.Run("returns only the requested conversation, oldest first", func(t *testing.T) {
-		storeResponse(t, "resp_idx_a1", "conv_idx_a", now)
-		storeResponse(t, "resp_idx_a2", "conv_idx_a", now+1)
-		storeResponse(t, "resp_idx_b1", "conv_idx_b", now)
-		// A response with no conversation must not land in any index.
-		storeResponse(t, "resp_idx_orphan", "", now)
-
-		responses, err := store.ListResponsesByConversation(ctx, "conv_idx_a", ListOptions{Order: "asc"})
-		require.NoError(t, err)
-		require.Len(t, responses, 2)
-		// The index is scored by created_at, so asc reads back chronologically.
-		assert.Equal(t, "resp_idx_a1", responses[0].ID)
-		assert.Equal(t, "resp_idx_a2", responses[1].ID)
-
-		assert.Equal(t, []string{"resp_idx_a1", "resp_idx_a2"}, conversationIndexMembers(t, store, "conv_idx_a"))
-		assert.Equal(t, []string{"resp_idx_b1"}, conversationIndexMembers(t, store, "conv_idx_b"))
-	})
-
-	t.Run("unknown conversation returns nothing", func(t *testing.T) {
-		responses, err := store.ListResponsesByConversation(ctx, "conv_idx_missing", ListOptions{})
-		assert.NoError(t, err)
-		assert.Empty(t, responses)
-	})
-
-	t.Run("empty conversation ID is rejected", func(t *testing.T) {
-		_, err := store.ListResponsesByConversation(ctx, "", ListOptions{})
-		assert.ErrorIs(t, err, ErrInvalidInput)
-	})
-
-	t.Run("deleting a response drops its index entry", func(t *testing.T) {
-		require.NoError(t, store.DeleteResponse(ctx, "resp_idx_a1"))
-
-		assert.Equal(t, []string{"resp_idx_a2"}, conversationIndexMembers(t, store, "conv_idx_a"))
-
-		responses, err := store.ListResponsesByConversation(ctx, "conv_idx_a", ListOptions{Order: "asc"})
-		require.NoError(t, err)
-		require.Len(t, responses, 1)
-		assert.Equal(t, "resp_idx_a2", responses[0].ID)
-	})
-
-	t.Run("moving a response between conversations reindexes it", func(t *testing.T) {
-		moved, err := store.GetResponse(ctx, "resp_idx_a2")
-		require.NoError(t, err)
-
-		moved.ConversationID = "conv_idx_b"
-		require.NoError(t, store.UpdateResponse(ctx, moved))
-
-		fromOld, err := store.ListResponsesByConversation(ctx, "conv_idx_a", ListOptions{Order: "asc"})
-		require.NoError(t, err)
-		assert.Empty(t, fromOld)
-		assert.Empty(t, conversationIndexMembers(t, store, "conv_idx_a"))
-
-		toNew, err := store.ListResponsesByConversation(ctx, "conv_idx_b", ListOptions{Order: "asc"})
-		require.NoError(t, err)
-		require.Len(t, toNew, 2)
-		assert.Equal(t, []string{"resp_idx_b1", "resp_idx_a2"}, []string{toNew[0].ID, toNew[1].ID})
-	})
-
-	t.Run("listing prunes entries whose payload is gone", func(t *testing.T) {
-		// Drop the payload behind the store's back, the way a TTL expiry would.
-		require.NoError(t, store.client.Del(ctx, store.buildKey(ResponseKeyPrefix+"resp_idx_b1")).Err())
-
-		responses, err := store.ListResponsesByConversation(ctx, "conv_idx_b", ListOptions{Order: "asc"})
-		require.NoError(t, err)
-		require.Len(t, responses, 1)
-		assert.Equal(t, "resp_idx_a2", responses[0].ID)
-
-		assert.Equal(t, []string{"resp_idx_a2"}, conversationIndexMembers(t, store, "conv_idx_b"))
-	})
-}
-
-// TestRedisDeleteConversationCascade uses more responses than one page: the
-// cascade reads the index directly, so it must not stop at DefaultListLimit.
-func TestRedisDeleteConversationCascade(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	const (
-		convID      = "conv_cascade"
-		responseCnt = DefaultListLimit + 5
-		otherConvID = "conv_cascade_other"
-		otherRespID = "resp_cascade_other"
-	)
-
-	require.NoError(t, store.CreateConversation(ctx, &responseapi.StoredConversation{
-		ID:        convID,
-		CreatedAt: time.Now().Unix(),
-	}))
-
-	for i := 0; i < responseCnt; i++ {
-		require.NoError(t, store.StoreResponse(ctx, &responseapi.StoredResponse{
-			ID:             fmt.Sprintf("resp_cascade_%d", i),
-			ConversationID: convID,
-			Status:         "completed",
-			CreatedAt:      time.Now().Unix() + int64(i),
-			Output:         []responseapi.OutputItem{{ID: fmt.Sprintf("item_%d", i)}},
-		}))
-	}
-
-	// A response in a different conversation must survive the cascade.
-	require.NoError(t, store.StoreResponse(ctx, &responseapi.StoredResponse{
-		ID:             otherRespID,
-		ConversationID: otherConvID,
-		Status:         "kept",
-		CreatedAt:      time.Now().Unix(),
-	}))
-
-	require.NoError(t, store.DeleteConversation(ctx, convID, true))
-
-	for i := 0; i < responseCnt; i++ {
-		_, err := store.GetResponse(ctx, fmt.Sprintf("resp_cascade_%d", i))
-		assert.ErrorIsf(t, err, ErrNotFound, "response %d should have been deleted with its conversation", i)
-	}
-	assert.Empty(t, conversationIndexMembers(t, store, convID))
-
-	survivor, err := store.GetResponse(ctx, otherRespID)
-	require.NoError(t, err)
-	assert.Equal(t, "kept", survivor.Status)
-}
-
-// TestRedisStoreResponseRejectsDuplicate covers the SET NX path: it enforces the
-// no-duplicate contract and orders the payload write ahead of the index write.
-func TestRedisStoreResponseRejectsDuplicate(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	response := &responseapi.StoredResponse{
-		ID:             "resp_duplicate",
-		ConversationID: "conv_duplicate",
-		Status:         "completed",
-		CreatedAt:      time.Now().Unix(),
-	}
-	require.NoError(t, store.StoreResponse(ctx, response))
-
-	duplicate := *response
-	duplicate.Status = "overwritten"
-	assert.ErrorIs(t, store.StoreResponse(ctx, &duplicate), ErrAlreadyExists)
-
-	// The original payload must be untouched, and indexed exactly once.
-	stored, err := store.GetResponse(ctx, response.ID)
-	require.NoError(t, err)
-	assert.Equal(t, "completed", stored.Status)
-	assert.Equal(t, []string{"resp_duplicate"}, conversationIndexMembers(t, store, "conv_duplicate"))
-}
-
-// directSetResponsePayload writes a response payload straight to Redis,
-// bypassing StoreResponse and its indexing entirely — the same shape as data
-// written before this index existed, or as an in-flight writer that landed
-// its payload but has not indexed it yet.
-func directSetResponsePayload(t *testing.T, store *RedisStore, response *responseapi.StoredResponse) {
-	t.Helper()
-
-	data, err := json.Marshal(response)
-	require.NoError(t, err)
-
-	key := store.buildKey(ResponseKeyPrefix + response.ID)
-	require.NoError(t, store.client.Set(context.Background(), key, data, store.ttl).Err())
-}
-
-// TestRedisStoreResponseRollsBackIndexFailure covers blueprint §2.3/§3.5: if
-// the payload write succeeds but indexing fails, StoreResponse must roll the
-// payload back via compare-delete rather than leave an orphan that a retry
-// can never repair (SETNX would just see it exists and hit ErrAlreadyExists
-// with no index to fix).
-func TestRedisStoreResponseRollsBackIndexFailure(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	injectedErr := errors.New("injected index failure")
-	store.indexResponseOverride = func(context.Context, string, string, int64) error {
-		return injectedErr
-	}
-
-	response := &responseapi.StoredResponse{
-		ID:             "resp_rollback",
-		ConversationID: "conv_rollback",
-		Status:         "completed",
-		CreatedAt:      time.Now().Unix(),
-	}
-
-	err := store.StoreResponse(ctx, response)
-	require.Error(t, err)
-	assert.ErrorIs(t, err, injectedErr)
-
-	// The rolled-back payload must be gone, not orphaned: GetResponse sees a
-	// clean slate, and a retry does not hit ErrAlreadyExists.
-	_, err = store.GetResponse(ctx, response.ID)
-	assert.ErrorIs(t, err, ErrNotFound)
-	assert.Empty(t, conversationIndexMembers(t, store, "conv_rollback"))
-
-	// Remove the injected failure and retry: must succeed cleanly.
-	store.indexResponseOverride = nil
-	require.NoError(t, store.StoreResponse(ctx, response))
-
-	stored, err := store.GetResponse(ctx, response.ID)
-	require.NoError(t, err)
-	assert.Equal(t, "completed", stored.Status)
-	assert.Equal(t, []string{"resp_rollback"}, conversationIndexMembers(t, store, "conv_rollback"))
-}
-
-// TestRedisCompareDeleteResponsePayload covers the ABA race the blueprint
-// calls out (§2.3): rollback must never delete a payload that changed after
-// the failed write, e.g. because a concurrent writer replaced it once the
-// original value's TTL expired.
-func TestRedisCompareDeleteResponsePayload(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	key := store.buildKey(ResponseKeyPrefix + "resp_cas")
-
-	t.Run("deletes when value matches", func(t *testing.T) {
-		require.NoError(t, store.client.Set(ctx, key, []byte("payload-a"), store.ttl).Err())
-
-		deleted, err := store.compareDeleteResponsePayload(ctx, key, []byte("payload-a"))
-		require.NoError(t, err)
-		assert.True(t, deleted)
-
-		exists, err := store.client.Exists(ctx, key).Result()
-		require.NoError(t, err)
-		assert.Zero(t, exists)
-	})
-
-	t.Run("leaves a changed value untouched", func(t *testing.T) {
-		require.NoError(t, store.client.Set(ctx, key, []byte("payload-a"), store.ttl).Err())
-
-		deleted, err := store.compareDeleteResponsePayload(ctx, key, []byte("payload-b"))
-		require.NoError(t, err)
-		assert.False(t, deleted)
-
-		value, err := store.client.Get(ctx, key).Result()
-		require.NoError(t, err)
-		assert.Equal(t, "payload-a", value)
-	})
-
-	t.Run("no-op against a missing key", func(t *testing.T) {
-		require.NoError(t, store.client.Del(ctx, key).Err())
-
-		deleted, err := store.compareDeleteResponsePayload(ctx, key, []byte("anything"))
-		require.NoError(t, err)
-		assert.False(t, deleted)
-	})
-}
-
-// TestRedisStoreResponseDuplicateRepairsOrphanedIndex covers the repair half
-// of blueprint §3.5: a retry that lands on an existing payload whose index
-// entry is missing (e.g. left behind by a prior partial failure, or never
-// indexed pre-upgrade) must repair the index from the stored payload before
-// returning the duplicate error.
-func TestRedisStoreResponseDuplicateRepairsOrphanedIndex(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	response := &responseapi.StoredResponse{
-		ID:             "resp_orphan",
-		ConversationID: "conv_orphan",
-		Status:         "completed",
-		CreatedAt:      time.Now().Unix(),
-	}
-	directSetResponsePayload(t, store, response)
-	require.Empty(t, conversationIndexMembers(t, store, "conv_orphan"))
-
-	retry := *response
-	err := store.StoreResponse(ctx, &retry)
-	assert.ErrorIs(t, err, ErrAlreadyExists)
-
-	assert.Equal(t, []string{"resp_orphan"}, conversationIndexMembers(t, store, "conv_orphan"))
-}
-
-// TestRedisStoreResponseDuplicateDifferentConversationNotRepaired covers the
-// poisoning guard from blueprint §2.3/§3.5: a duplicate ID whose stored
-// payload belongs to a different conversation than the retry attempted must
-// not be indexed under the attempted conversation.
-func TestRedisStoreResponseDuplicateDifferentConversationNotRepaired(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	stored := &responseapi.StoredResponse{
-		ID:             "resp_mismatch",
-		ConversationID: "conv_real",
-		Status:         "completed",
-		CreatedAt:      time.Now().Unix(),
-	}
-	directSetResponsePayload(t, store, stored)
-
-	attempted := *stored
-	attempted.ConversationID = "conv_wrong"
-	err := store.StoreResponse(ctx, &attempted)
-	assert.ErrorIs(t, err, ErrAlreadyExists)
-
-	assert.Empty(t, conversationIndexMembers(t, store, "conv_wrong"))
-	// The real conversation was never indexed either: repair only runs
-	// against the attempted conversation, and stored != attempted here.
-	assert.Empty(t, conversationIndexMembers(t, store, "conv_real"))
-}
-
-// TestRedisStoreResponseDuplicateNoConversationNotRepaired covers the "no
-// index expected" branch of repairExistingResponseIndex: a stored response
-// with no ConversationID has nothing to repair.
-func TestRedisStoreResponseDuplicateNoConversationNotRepaired(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	stored := &responseapi.StoredResponse{
-		ID:        "resp_no_conv",
-		Status:    "completed",
-		CreatedAt: time.Now().Unix(),
-	}
-	directSetResponsePayload(t, store, stored)
-
-	attempted := *stored
-	attempted.ConversationID = "conv_should_not_exist"
-	err := store.StoreResponse(ctx, &attempted)
-	assert.ErrorIs(t, err, ErrAlreadyExists)
-
-	assert.Empty(t, conversationIndexMembers(t, store, "conv_should_not_exist"))
-}
-
-// TestRedisListResponsesByConversationLazyBackfill covers the upgrade path
-// (blueprint §6.2): responses written before the index existed — simulated
-// with directSetResponsePayload, which bypasses StoreResponse's indexing —
-// must still be discoverable on first read, get backfilled into the index,
-// and must not trigger a second O(N) scan on the next read of the same
-// conversation.
-func TestRedisListResponsesByConversationLazyBackfill(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	now := time.Now().Unix()
-	legacy := []*responseapi.StoredResponse{
-		{ID: "resp_legacy_1", ConversationID: "conv_legacy", Status: "completed", CreatedAt: now},
-		{ID: "resp_legacy_2", ConversationID: "conv_legacy", Status: "completed", CreatedAt: now + 1},
-		{ID: "resp_legacy_3", ConversationID: "conv_legacy", Status: "completed", CreatedAt: now + 2},
-	}
-	for _, response := range legacy {
-		directSetResponsePayload(t, store, response)
-	}
-	require.Empty(t, conversationIndexMembers(t, store, "conv_legacy"),
-		"precondition: no index should exist yet for legacy data")
-
-	responses, err := store.ListResponsesByConversation(ctx, "conv_legacy", ListOptions{Order: "asc"})
-	require.NoError(t, err)
-	require.Len(t, responses, 3)
-	assert.Equal(t, []string{"resp_legacy_1", "resp_legacy_2", "resp_legacy_3"},
-		[]string{responses[0].ID, responses[1].ID, responses[2].ID})
-	assert.Equal(t, 1, store.scanInvocations, "first read must run exactly one legacy scan")
-
-	assert.ElementsMatch(t, []string{"resp_legacy_1", "resp_legacy_2", "resp_legacy_3"},
-		conversationIndexMembers(t, store, "conv_legacy"))
-	assert.Zero(t, exists(t, store, store.emptyConversationIndexMarkerKey("conv_legacy")))
-
-	// The index now exists, so a second read must go straight through it
-	// without scanning the keyspace again.
-	responses, err = store.ListResponsesByConversation(ctx, "conv_legacy", ListOptions{Order: "asc"})
-	require.NoError(t, err)
-	require.Len(t, responses, 3)
-	assert.Equal(t, 1, store.scanInvocations, "second read of an already-backfilled conversation must not scan again")
-}
-
-// exists reports whether a raw Redis key exists, for asserting on marker/
-// index presence directly rather than only through store methods.
-func exists(t *testing.T, store *RedisStore, key string) int64 {
-	t.Helper()
-	n, err := store.client.Exists(context.Background(), key).Result()
-	require.NoError(t, err)
-	return n
-}
-
-// TestRedisListResponsesByConversationEmptyMarker covers blueprint §2.4/§3.4:
-// a legitimately empty or unknown conversation must not force a full
-// keyspace scan on every read — only once, after which the empty marker
-// short-circuits subsequent reads.
-func TestRedisListResponsesByConversationEmptyMarker(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	// Seed unrelated data so the scan has something to walk past.
-	directSetResponsePayload(t, store, &responseapi.StoredResponse{
-		ID: "resp_unrelated", ConversationID: "conv_other", Status: "completed", CreatedAt: time.Now().Unix(),
-	})
-
-	responses, err := store.ListResponsesByConversation(ctx, "conv_never_existed", ListOptions{})
-	require.NoError(t, err)
-	assert.Empty(t, responses)
-	assert.Equal(t, 1, store.scanInvocations)
-
-	markerKey := store.emptyConversationIndexMarkerKey("conv_never_existed")
-	assert.EqualValues(t, 1, exists(t, store, markerKey))
-	ttl, err := store.client.TTL(ctx, markerKey).Result()
-	require.NoError(t, err)
-	assert.Positive(t, ttl, "empty marker must carry a positive TTL, not be immortal")
-
-	// Repeated reads of the same empty conversation must not scan again.
-	for i := 0; i < 3; i++ {
-		responses, err = store.ListResponsesByConversation(ctx, "conv_never_existed", ListOptions{})
-		require.NoError(t, err)
-		assert.Empty(t, responses)
-	}
-	assert.Equal(t, 1, store.scanInvocations, "repeated reads of an empty conversation must not force repeated scans")
-}
-
-// TestRedisLazyBackfillConcurrentWriteNotHidden covers blueprint §2.2: a
-// response indexed normally by a concurrent StoreResponse call, landing
-// while a lazy legacy scan for the same conversation is in flight, must
-// survive — the scan's own findings only ever add to the index, and
-// indexResponse always clears the empty marker, so neither ordering can make
-// the concurrent write disappear behind a "confirmed empty" marker.
-func TestRedisLazyBackfillConcurrentWriteNotHidden(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	legacy := &responseapi.StoredResponse{
-		ID: "resp_race_legacy", ConversationID: "conv_race", Status: "completed", CreatedAt: time.Now().Unix(),
-	}
-	directSetResponsePayload(t, store, legacy)
-
-	concurrent := &responseapi.StoredResponse{
-		ID: "resp_race_concurrent", ConversationID: "conv_race", Status: "completed", CreatedAt: time.Now().Unix() + 1,
-	}
-	store.lazyBackfillPreScanHook = func() {
-		// Runs once, before the scan walks the keyspace: lands a normal
-		// indexed write for the same conversation the backfill is about to
-		// scan for, so the scan observes both the legacy and the
-		// concurrently-indexed payload.
-		require.NoError(t, store.StoreResponse(ctx, concurrent))
-	}
-
-	responses, err := store.ListResponsesByConversation(ctx, "conv_race", ListOptions{Order: "asc"})
-	require.NoError(t, err)
-	require.Len(t, responses, 2)
-	assert.Equal(t, []string{"resp_race_legacy", "resp_race_concurrent"},
-		[]string{responses[0].ID, responses[1].ID})
-
-	assert.ElementsMatch(t, []string{"resp_race_legacy", "resp_race_concurrent"},
-		conversationIndexMembers(t, store, "conv_race"))
-	assert.Zero(t, exists(t, store, store.emptyConversationIndexMarkerKey("conv_race")),
-		"the concurrently indexed write must clear any empty marker, not be hidden behind one")
-}
-
-// TestNormalizeResponseListOptions needs no Redis: pure validation/defaulting.
-func TestNormalizeResponseListOptions(t *testing.T) {
-	tests := []struct {
-		name      string
-		opts      ListOptions
-		want      normalizedListOptions
-		expectErr error
-	}{
-		{
-			name: "zero value defaults to desc and DefaultListLimit",
-			opts: ListOptions{},
-			want: normalizedListOptions{Limit: DefaultListLimit, Order: "desc"},
-		},
-		{
-			name: "negative limit defaults",
-			opts: ListOptions{Limit: -5},
-			want: normalizedListOptions{Limit: DefaultListLimit, Order: "desc"},
-		},
-		{
-			name: "limit above MaxListLimit is clamped",
-			opts: ListOptions{Limit: MaxListLimit + 500},
-			want: normalizedListOptions{Limit: MaxListLimit, Order: "desc"},
-		},
-		{
-			name: "explicit asc is kept",
-			opts: ListOptions{Order: "asc"},
-			want: normalizedListOptions{Limit: DefaultListLimit, Order: "asc"},
-		},
-		{
-			name: "explicit desc is kept",
-			opts: ListOptions{Order: "desc"},
-			want: normalizedListOptions{Limit: DefaultListLimit, Order: "desc"},
-		},
-		{
-			name: "after cursor is preserved",
-			opts: ListOptions{After: "resp_a"},
-			want: normalizedListOptions{Limit: DefaultListLimit, Order: "desc", After: "resp_a"},
-		},
-		{
-			name: "before cursor is preserved",
-			opts: ListOptions{Before: "resp_b"},
-			want: normalizedListOptions{Limit: DefaultListLimit, Order: "desc", Before: "resp_b"},
-		},
-		{
-			name:      "unrecognized order is rejected",
-			opts:      ListOptions{Order: "sideways"},
-			expectErr: ErrInvalidInput,
-		},
-		{
-			name:      "after and before together are rejected",
-			opts:      ListOptions{After: "resp_a", Before: "resp_b"},
-			expectErr: ErrInvalidInput,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := normalizeResponseListOptions(tt.opts)
-			if tt.expectErr != nil {
-				assert.ErrorIs(t, err, tt.expectErr)
-				return
-			}
-			require.NoError(t, err)
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
-// seedPageResponses stores count responses in convID with strictly
-// increasing CreatedAt (now+i) and IDs "resp_page_<i>", so ascending index
-// order is resp_page_0..resp_page_(count-1) and descending is the reverse.
-func seedPageResponses(t *testing.T, store *RedisStore, convID string, count int) []string {
-	t.Helper()
-	ctx := context.Background()
-	now := time.Now().Unix()
-
-	ids := make([]string, count)
-	for i := 0; i < count; i++ {
-		id := fmt.Sprintf("resp_page_%d", i)
-		ids[i] = id
-		require.NoError(t, store.StoreResponse(ctx, &responseapi.StoredResponse{
-			ID:             id,
-			ConversationID: convID,
-			Status:         "completed",
-			CreatedAt:      now + int64(i),
-		}))
-	}
-	return ids
-}
-
-func responseIDsOf(responses []*responseapi.StoredResponse) []string {
-	ids := make([]string, len(responses))
-	for i, r := range responses {
-		ids[i] = r.ID
-	}
-	return ids
-}
-
-// TestRedisListDefaultOrderIsDescending is the direct regression test for the
-// Phase 4 default-order fix: an unspecified Order must return newest first,
-// per the ListOptions.Order contract in interface.go (blueprint §3.6),
-// which neither store honored before this issue.
-func TestRedisListDefaultOrderIsDescending(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	seedPageResponses(t, store, "conv_default_order", 3)
-
-	responses, err := store.ListResponsesByConversation(ctx, "conv_default_order", ListOptions{})
-	require.NoError(t, err)
-	require.Len(t, responses, 3)
-	assert.Equal(t, []string{"resp_page_2", "resp_page_1", "resp_page_0"}, responseIDsOf(responses))
-}
-
-// TestRedisListBoundedLimit covers blueprint §6.5: default and clamped
-// limits, without reading the full conversation.
-func TestRedisListBoundedLimit(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	total := DefaultListLimit + 5
-	seedPageResponses(t, store, "conv_bounded", total)
-
-	t.Run("default limit caps at DefaultListLimit, not the full conversation", func(t *testing.T) {
-		responses, err := store.ListResponsesByConversation(ctx, "conv_bounded", ListOptions{Order: "asc"})
-		require.NoError(t, err)
-		assert.Len(t, responses, DefaultListLimit)
-		// Bounded read: the oldest page, not an arbitrary DefaultListLimit
-		// subset of a fully-read-then-truncated slice.
-		assert.Equal(t, "resp_page_0", responses[0].ID)
-	})
-
-	t.Run("limit above MaxListLimit is clamped", func(t *testing.T) {
-		responses, err := store.ListResponsesByConversation(ctx, "conv_bounded", ListOptions{Order: "asc", Limit: MaxListLimit + 1000})
-		require.NoError(t, err)
-		assert.LessOrEqual(t, len(responses), MaxListLimit)
-	})
-}
-
-// TestRedisListCursors covers blueprint §3.6/§6.5's After/Before rank-window
-// semantics in both directions, plus the invalid-input and unknown-cursor
-// edge cases.
-func TestRedisListCursors(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	seedPageResponses(t, store, "conv_cursor", 5) // resp_page_0 (oldest) .. resp_page_4 (newest)
-
-	list := func(t *testing.T, opts ListOptions) []string {
-		t.Helper()
-		responses, err := store.ListResponsesByConversation(ctx, "conv_cursor", opts)
-		require.NoError(t, err)
-		return responseIDsOf(responses)
-	}
-
-	t.Run("ascending pages walk oldest to newest via After", func(t *testing.T) {
-		assert.Equal(t, []string{"resp_page_0", "resp_page_1"}, list(t, ListOptions{Order: "asc", Limit: 2}))
-		assert.Equal(t, []string{"resp_page_2", "resp_page_3"},
-			list(t, ListOptions{Order: "asc", Limit: 2, After: "resp_page_1"}))
-		assert.Equal(t, []string{"resp_page_4"},
-			list(t, ListOptions{Order: "asc", Limit: 2, After: "resp_page_3"}))
-		assert.Empty(t, list(t, ListOptions{Order: "asc", Limit: 2, After: "resp_page_4"}))
-	})
-
-	t.Run("descending pages walk newest to oldest via After", func(t *testing.T) {
-		assert.Equal(t, []string{"resp_page_4", "resp_page_3"}, list(t, ListOptions{Order: "desc", Limit: 2}))
-		assert.Equal(t, []string{"resp_page_2", "resp_page_1"},
-			list(t, ListOptions{Order: "desc", Limit: 2, After: "resp_page_3"}))
-	})
-
-	t.Run("before returns the page preceding the cursor in the selected order", func(t *testing.T) {
-		assert.Equal(t, []string{"resp_page_1", "resp_page_2"},
-			list(t, ListOptions{Order: "asc", Limit: 2, Before: "resp_page_3"}))
-		// Nothing precedes the very first element.
-		assert.Empty(t, list(t, ListOptions{Order: "asc", Limit: 2, Before: "resp_page_0"}))
-	})
-
-	t.Run("a cursor naming a non-member returns an empty page, not an error", func(t *testing.T) {
-		assert.Empty(t, list(t, ListOptions{Order: "asc", After: "resp_does_not_exist"}))
-		assert.Empty(t, list(t, ListOptions{Order: "asc", Before: "resp_does_not_exist"}))
-	})
-
-	t.Run("after and before together are rejected", func(t *testing.T) {
-		_, err := store.ListResponsesByConversation(ctx, "conv_cursor", ListOptions{After: "resp_page_0", Before: "resp_page_4"})
-		assert.ErrorIs(t, err, ErrInvalidInput)
-	})
-
-	t.Run("unrecognized order is rejected", func(t *testing.T) {
-		_, err := store.ListResponsesByConversation(ctx, "conv_cursor", ListOptions{Order: "sideways"})
-		assert.ErrorIs(t, err, ErrInvalidInput)
-	})
-}
-
-// TestRedisAddResponseToConversation covers blueprint §3.7/§5 Phase 5: the
-// stored response's own ConversationID is the only source of truth for
-// whether it may be indexed under conversationID.
-func TestRedisAddResponseToConversation(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	t.Run("stored response matching the conversation is indexed", func(t *testing.T) {
-		directSetResponsePayload(t, store, &responseapi.StoredResponse{
-			ID: "resp_add_match", ConversationID: "conv_add", Status: "completed", CreatedAt: time.Now().Unix(),
-		})
-		require.Empty(t, conversationIndexMembers(t, store, "conv_add"))
-
-		require.NoError(t, store.AddResponseToConversation(ctx, "conv_add", "resp_add_match"))
-		assert.Equal(t, []string{"resp_add_match"}, conversationIndexMembers(t, store, "conv_add"))
-	})
-
-	t.Run("missing response is ErrNotFound", func(t *testing.T) {
-		err := store.AddResponseToConversation(ctx, "conv_add", "resp_add_missing")
-		assert.ErrorIs(t, err, ErrNotFound)
-	})
-
-	t.Run("response belonging to a different conversation is rejected", func(t *testing.T) {
-		directSetResponsePayload(t, store, &responseapi.StoredResponse{
-			ID: "resp_add_other", ConversationID: "conv_add_real", Status: "completed", CreatedAt: time.Now().Unix(),
-		})
-
-		err := store.AddResponseToConversation(ctx, "conv_add_wrong", "resp_add_other")
-		assert.ErrorIs(t, err, ErrInvalidInput)
-		assert.Empty(t, conversationIndexMembers(t, store, "conv_add_wrong"))
-	})
-
-	t.Run("response with no stored conversation is rejected", func(t *testing.T) {
-		directSetResponsePayload(t, store, &responseapi.StoredResponse{
-			ID: "resp_add_no_conv", Status: "completed", CreatedAt: time.Now().Unix(),
-		})
-
-		err := store.AddResponseToConversation(ctx, "conv_add_none", "resp_add_no_conv")
-		assert.ErrorIs(t, err, ErrInvalidInput)
-		assert.Empty(t, conversationIndexMembers(t, store, "conv_add_none"))
-	})
-}
-
-// TestRedisUpdateResponseRestoresOnIndexFailure covers the UpdateResponse
-// repairability the blueprint asks for on top of StoreResponse's (§5 Phase
-// 5): if the new conversation's index write fails, the previous payload
-// bytes must be restored and reindexed under the previous conversation,
-// rather than left pointing at a conversation whose index was never
-// actually written.
-func TestRedisUpdateResponseRestoresOnIndexFailure(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	original := &responseapi.StoredResponse{
-		ID:             "resp_update_rollback",
-		ConversationID: "conv_update_from",
-		Status:         "original",
-		CreatedAt:      time.Now().Unix(),
-	}
-	require.NoError(t, store.StoreResponse(ctx, original))
-	require.Equal(t, []string{"resp_update_rollback"}, conversationIndexMembers(t, store, "conv_update_from"))
-
-	injectedErr := errors.New("injected update index failure")
-	store.indexResponseOverride = func(context.Context, string, string, int64) error {
-		return injectedErr
-	}
-
-	updated := *original
-	updated.ConversationID = "conv_update_to"
-	updated.Status = "updated"
-	err := store.UpdateResponse(ctx, &updated)
-	require.Error(t, err)
-	assert.ErrorIs(t, err, injectedErr)
-
-	// The payload restore itself does not go through indexResponseOverride
-	// (it's a plain SET), so it must have succeeded even with the override
-	// still failing indexResponse.
-	store.indexResponseOverride = nil
-
-	restored, err := store.GetResponse(ctx, original.ID)
-	require.NoError(t, err)
-	assert.Equal(t, "original", restored.Status, "payload must be restored to its pre-update value")
-	assert.Equal(t, "conv_update_from", restored.ConversationID)
-
-	assert.Empty(t, conversationIndexMembers(t, store, "conv_update_to"),
-		"the failed new conversation must never gain an index entry")
-	assert.Equal(t, []string{"resp_update_rollback"}, conversationIndexMembers(t, store, "conv_update_from"),
-		"the previous conversation must be reindexed after rollback")
-}
-
-// TestRedisDeleteConversationCascadeBatched exceeds redisDeleteBatchSize so
-// the cascade must loop more than once (blueprint §5 Phase 5 / §6.6).
-func TestRedisDeleteConversationCascadeBatched(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	const (
-		convID      = "conv_cascade_batched"
-		responseCnt = redisDeleteBatchSize + 50
-	)
-
-	require.NoError(t, store.CreateConversation(ctx, &responseapi.StoredConversation{
-		ID: convID, CreatedAt: time.Now().Unix(),
-	}))
-	for i := 0; i < responseCnt; i++ {
-		require.NoError(t, store.StoreResponse(ctx, &responseapi.StoredResponse{
-			ID:             fmt.Sprintf("resp_cascade_batched_%d", i),
-			ConversationID: convID,
-			Status:         "completed",
-			CreatedAt:      time.Now().Unix() + int64(i),
-		}))
-	}
-
-	require.NoError(t, store.DeleteConversation(ctx, convID, true))
-
-	for i := 0; i < responseCnt; i++ {
-		_, err := store.GetResponse(ctx, fmt.Sprintf("resp_cascade_batched_%d", i))
-		assert.ErrorIsf(t, err, ErrNotFound, "response %d should have been deleted across batches", i)
-	}
-	assert.Empty(t, conversationIndexMembers(t, store, convID))
-	assert.Zero(t, exists(t, store, store.emptyConversationIndexMarkerKey(convID)))
-	_, err := store.GetConversation(ctx, convID)
-	assert.ErrorIs(t, err, ErrNotFound)
-}
-
-// TestRedisDeleteConversationCascadeFailureLeavesConversation covers
-// blueprint §5 Phase 5: a cascade failure must be reported to the caller,
-// and the conversation record must survive as the retry anchor rather than
-// be deleted ahead of (and regardless of) the cascade's outcome.
-func TestRedisDeleteConversationCascadeFailureLeavesConversation(t *testing.T) {
-	store := newConversationIndexStore(t)
-	ctx := context.Background()
-
-	const convID = "conv_cascade_failure"
-	require.NoError(t, store.CreateConversation(ctx, &responseapi.StoredConversation{
-		ID: convID, CreatedAt: time.Now().Unix(),
-	}))
-	require.NoError(t, store.StoreResponse(ctx, &responseapi.StoredResponse{
-		ID: "resp_cascade_failure", ConversationID: convID, Status: "completed", CreatedAt: time.Now().Unix(),
-	}))
-
-	injectedErr := errors.New("injected cascade delete failure")
-	store.deleteResponseBatchOverride = func(context.Context, []string) error {
-		return injectedErr
-	}
-
-	err := store.DeleteConversation(ctx, convID, true)
-	require.Error(t, err)
-	assert.ErrorIs(t, err, injectedErr)
-
-	// The conversation record must still be there: the caller can retry.
-	_, getErr := store.GetConversation(ctx, convID)
-	assert.NoError(t, getErr)
-	// The response and its index entry are untouched by the failed batch.
-	_, getErr = store.GetResponse(ctx, "resp_cascade_failure")
-	assert.NoError(t, getErr)
-	assert.Equal(t, []string{"resp_cascade_failure"}, conversationIndexMembers(t, store, convID))
-
-	store.deleteResponseBatchOverride = nil
-	require.NoError(t, store.DeleteConversation(ctx, convID, true))
-	_, getErr = store.GetConversation(ctx, convID)
-	assert.ErrorIs(t, getErr, ErrNotFound)
 }
